Fix warning C4838
BUG=#315
TEST=
REF_BUG=26932823
REF_CL=114656293
REF_TIME=2016-02-14T17:45:29-08:00
REF_TIME_RAW=1455500729 -0800

+++ src/renderer/win32/text_renderer.cc
@@ -400,8 +400,10 @@ class DirectWriteTextRenderer : public TextRenderer {
     for (size_t i = 0; i < display_list.size(); ++i) {
       const auto &item = display_list[i];
       const D2D1_RECT_F render_rect = {
-        item.rect.Left(), item.rect.Top(), item.rect.Right(),
-        item.rect.Bottom(),
+        static_cast<float>(item.rect.Left()),
+        static_cast<float>(item.rect.Top()),
+        static_cast<float>(item.rect.Right()),
+        static_cast<float>(item.rect.Bottom()),
       };
       dc_render_target_->DrawText(item.text.data(),
                                   item.text.size(),
